Don't depend on the abandoned docker module

+++ manifests/docker_run.pp
@@ -24,16 +24,27 @@ define sunet::docker_run(
   Boolean $uid_gid_consistency = true,
   Boolean $fetch_docker_image  = true,
 ) {
-  if $::facts['sunet_nftables_enabled'] == 'yes' {
-    if ! has_key($::facts['networking']['interfaces'], 'to_docker') {
-      notice("sunet::docker_compose: No to_docker interface found, not installing ${name}")
-      $_install_service = false
+
+  $docker_class = $::facts['dockerhost2'] ? {
+    no => 'sunet::dockerhost',
+    default => 'sunet::dockerhost2',
+  }
+
+  if ($docker_class == 'sunet::dockerhost') {
+    if $::facts['sunet_nftables_enabled'] == 'yes' {
+      if ! has_key($::facts['networking']['interfaces'], 'to_docker') {
+        notice("sunet::docker_compose: No to_docker interface found, not installing ${name}")
+        $_install_service = false
+      } else {
+        $_install_service = true
+      }
     } else {
       $_install_service = true
     }
   } else {
     $_install_service = true
   }
+
 
   if $_install_service {
     if $use_unbound {
@@ -61,6 +72,7 @@ define sunet::docker_run(
     }
 
     $image_tag = "${image}:${imagetag}"
+
     if $fetch_docker_image {
       docker::image { "${name}_${image_tag}" :  # make it possible to use the same docker image more than once on a node
         ensure  => $ensure,
@@ -69,32 +81,49 @@ define sunet::docker_run(
       }
     }
 
-    docker::run { $name :
-      ensure                   => $ensure,
-      volumes                  => flatten([$volumes, $_uid_gid]),
-      hostname                 => $hostname,
-      ports                    => $ports,
-      expose                   => $expose,
-      env                      => $env,
-      net                      => $net,
-      command                  => $command,
-      dns                      => $dns,
-      depends                  => $_depends,
-      require                  => flatten([$req]),
-      before_start             => $before_start,
-      before_stop              => $before_stop,
-      after_start              => $after_start,
-      after_stop               => $after_stop,
-      docker_service           => true,  # the service 'docker' is maintainer by puppet, so depend on it
-      image                    => $image_tag,
-      extra_parameters         => flatten([$extra_parameters]),
-      extra_systemd_parameters => $extra_systemd_parameters,
-    }
+    if ($docker_class == 'sunet::dockerhost') {
+      docker::run { $name :
+        ensure                   => $ensure,
+        volumes                  => flatten([$volumes, $_uid_gid]),
+        hostname                 => $hostname,
+        ports                    => $ports,
+        expose                   => $expose,
+        env                      => $env,
+        net                      => $net,
+        command                  => $command,
+        dns                      => $dns,
+        depends                  => $_depends,
+        require                  => flatten([$req]),
+        before_start             => $before_start,
+        before_stop              => $before_stop,
+        after_start              => $after_start,
+        after_stop               => $after_stop,
+        docker_service           => true,  # the service 'docker' is maintainer by puppet, so depend on it
+        image                    => $image_tag,
+        extra_parameters         => flatten([$extra_parameters]),
+        extra_systemd_parameters => $extra_systemd_parameters,
+      }
 
-    # Remove the upstart file created by earlier versions of garethr-docker
-    exec { "remove_docker_upstart_job_${name}":
-      command => "/bin/rm /etc/init/docker-${name}.conf",
-      onlyif  => "/usr/bin/test -f /etc/init/docker-${name}.conf",
+      # Remove the upstart file created by earlier versions of garethr-docker
+      exec { "remove_docker_upstart_job_${name}":
+        command => "/bin/rm /etc/init/docker-${name}.conf",
+        onlyif  => "/usr/bin/test -f /etc/init/docker-${name}.conf",
+      }
+    } else {
+
+      service { "docker-${name}":
+        ensure => 'stopped',
+        enable =>  'false',
+      }
+
+      $flat_volumes = flatten([$volumes, $_uid_gid])
+      sunet::docker_compose { $name:
+        content          => template('sunet/docker/docker-compose.yml.erb'),
+        service_name     => $name,
+        compose_dir      => "/opt/docker_run/${name}",
+        compose_filename => 'docker-compose.yml',
+        description      => "${name} shrimed fom docker_run",
+      }
     }
   }
 }
